notification-service/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
python-jose[cryptography]>=3.3.0
aiosqlite>=0.20.0
httpx>=0.28.0
websockets>=14.0
